version 2025.04.27.2.44.p
so, right now what is working
- api rest deliver => deliver app mobil (authentification, getCurentUser)
- api rest admin => admin plateform (authetificate, getCurentUser)

- websocket deliver => deliveryApp (next to implement some auto emit like send position each 5 seconds or notifications, etc ...)

-websock et admin, ready also bu not yet used

- models deliver, admin, cand of ok for the moment

admin with some role but in two groups 'admin' and 'pharmacits' => to implement complet with example

Bref : starting to looks nice and complexe. intelligent server that toogle between two firebase apps.
Question : i'll support bigs demands, trafics and simultaneous comunications ?

may be replicat server in function of number of user connected (like when more than 1000 users connected, create a new server or something like)

so transform the initial server as a distributor that will just get call, transfer to the real server that will make the stuffs, and send back the answer (bof maybe in long time after)

othe interogation, is mongo will be strong enought, when datas will be bigs as fuck ? and server will be able to be speed also.